refactor: 의존 역전 현상 변경
