+++ src/main/java/blackjack/domain/state/Running.java
@@ -1,22 +1,11 @@
 package blackjack.domain.state;
 
-import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 
 public abstract class Running extends Started {
 
     protected Running(Cards cards) {
         super(cards);
-    }
-
-    @Override
-    public final State draw(Card card) {
-        final Cards cards = getCards();
-        cards.receiveCard(card);
-        if (cards.isBust()) {
-            return new Bust(cards);
-        }
-        return new Hit(cards);
     }
 
     @Override
